+++ geopackage-sdk/src/androidTest/java/mil/nga/geopackage/GeoPackageImportTest.java
@@ -121,4 +121,17 @@ public class GeoPackageImportTest extends ImportGeoPackageTestCase {
 
 	}
 
+	/**
+	 * Test user daos
+	 *
+	 * @throws SQLException
+	 *             upon error
+	 */
+	@Test
+	public void testUserDao() throws SQLException {
+
+		GeoPackageTestUtils.testUserDao(geoPackage);
+
+	}
+
 }