+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 * geopackage-core version 6.6.7
 * Ignored internal databases support for non GeoPackages (preconfigured to ignore ue3.db for Google Maps)
+* Get abstract User DAO by table name
+* Set User Custom DAO contents
 
 ## [6.7.3](https://github.com/ngageoint/geopackage-android/releases/tag/6.7.3) (11-30-2023)